Starts adding V1 unit tests
